refactor(analysis): move the engine + use cases into the vailabel-analysis crate
Continue emptying src/domain: pull the Dataset Intelligence logic out of
`src/domain/analysis` into the `vailabel-analysis` crate, leaving the binary
with only the background-job lifecycle (worker thread + the `analysis://progress`
Tauri event).

- domain: the pure `engine` (metadata analysis, geometry, outlier stats, and the
  report-assembly helpers folded in from the old service) + the
  `AnalysisRepository` port. No I/O.
- application: `AnalysisAppService` (the run pipeline + report CRUD) plus the
  `ImageDecoder` (pixel pass) and `AnalysisReporter` (progress) ports.
- infrastructure: `ImageQualityDecoder` — the `image`-crate pixel decode +
  blur/exposure/embedding metrics, behind the port (the only layer reading files).
- binary: `AnalysisStoreRepository` adapts the port over the residual
  `DesktopStore` (source rows via `list_by_field`, reports via the
  `analysis_reports` table — logic-only, no schema surgery); the facade keeps the
  job map, worker thread, and progress emit, driving the crate via
  `AnalysisReporter`.
- arch-tests: `analysis` moves FULLY_PURE -> LAYERED (its infra is the image
  decoder). Also fixed a scanner false-positive — `ort::` matched the tail of
  `ImageQualityReport::`; the matcher now requires a word boundary before the
  token (with a regression assertion).

IPC commands and `analysis://progress` payloads are unchanged. Green:
`cargo build -p vailabel-analysis`, arch-tests (4), default `cargo check`.
