+++ apps/studio/src-tauri/crates/analysis/Cargo.toml
@@ -2,11 +2,16 @@
 name = "vailabel-analysis"
 version = "0.1.0"
 edition = "2021"
-description = "Dataset Intelligence module: the analysis report model (health, analytics, quality, image-quality, outliers, findings) and request DTOs. A dataset-statistics/health concern; pure domain types."
+description = "Dataset Intelligence module: the analysis report model (health, analytics, quality, image-quality, outliers, findings), the pure metadata/outlier engine, and the analysis use cases. Domain/application/contracts are pure; image-pixel decoding lives behind the ImageDecoder port in infrastructure."
 
 [lib]
 name = "vailabel_analysis"
 
 [dependencies]
 vailabel-core = { workspace = true }
+vailabel-shared = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
+# Infrastructure layer only: decode image pixels for the blur/exposure/embedding
+# metrics, behind the `ImageDecoder` port.
+image = { version = "0.25", default-features = false, features = ["jpeg", "png", "gif"] }
