+++ apps/studio/src-tauri/src/domain/analysis/mod.rs
@@ -1,4 +1,4 @@
 pub mod commands;
-pub mod engine;
 pub mod model;
+pub mod repository;
 pub mod service;
